requirements.txt
fastapi
uvicorn
pydantic
pyyaml
httpx
pytest
numpy
rank-bm25
sentence-transformers
sentence-transformers
rank-bm25
numpy
